Remove usage of `itertools` (#15)
Itertools crate was used for ergonomic join on the iterator of strings.
Changed this to use join on the vec of strings from the std lib, and
droped a dependecy of the crate. Std lib join should even be more
performant, since it prealocates the string length correctly, while
itertools is doing the best effort it can.